refactor: clean up entrypoints
